Move `StyleProperty` to `styled_text`
