+++ styled_text/src/lib.rs
@@ -27,6 +27,7 @@ mod font_family;
 mod font_settings;
 mod font_stack;
 mod generic;
+mod style_property;
 
 pub use attributes::{Stretch, Style, Weight};
 pub use brush::Brush;
@@ -34,3 +35,4 @@ pub use font_family::FontFamily;
 pub use font_settings::{FontFeature, FontSettings, FontVariation};
 pub use font_stack::FontStack;
 pub use generic::GenericFamily;
+pub use style_property::StyleProperty;
