+++ styled_text/src/style_property.rs
@@ -1,0 +1,76 @@
+// Copyright 2021 the Parley Authors
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+
+use crate::{
+    Brush, FontFamily, FontFeature, FontSettings, FontStack, FontVariation, GenericFamily,
+    Stretch as FontStretch, Style as FontStyle, Weight as FontWeight,
+};
+
+/// Properties that define a style.
+#[derive(Clone, PartialEq, Debug)]
+pub enum StyleProperty<'a, B: Brush> {
+    /// Font family stack.
+    FontStack(FontStack<'a>),
+    /// Font size.
+    FontSize(f32),
+    /// Font stretch.
+    FontStretch(FontStretch),
+    /// Font style.
+    FontStyle(FontStyle),
+    /// Font weight.
+    FontWeight(FontWeight),
+    /// Font variation settings.
+    FontVariations(FontSettings<'a, FontVariation>),
+    /// Font feature settings.
+    FontFeatures(FontSettings<'a, FontFeature>),
+    /// Locale.
+    Locale(Option<&'a str>),
+    /// Brush for rendering text.
+    Brush(B),
+    /// Underline decoration.
+    Underline(bool),
+    /// Offset of the underline decoration.
+    UnderlineOffset(Option<f32>),
+    /// Size of the underline decoration.
+    UnderlineSize(Option<f32>),
+    /// Brush for rendering the underline decoration.
+    UnderlineBrush(Option<B>),
+    /// Strikethrough decoration.
+    Strikethrough(bool),
+    /// Offset of the strikethrough decoration.
+    StrikethroughOffset(Option<f32>),
+    /// Size of the strikethrough decoration.
+    StrikethroughSize(Option<f32>),
+    /// Brush for rendering the strikethrough decoration.
+    StrikethroughBrush(Option<B>),
+    /// Line height multiplier.
+    LineHeight(f32),
+    /// Extra spacing between words.
+    WordSpacing(f32),
+    /// Extra spacing between letters.
+    LetterSpacing(f32),
+}
+
+impl<'a, B: Brush> From<FontStack<'a>> for StyleProperty<'a, B> {
+    fn from(fs: FontStack<'a>) -> Self {
+        StyleProperty::FontStack(fs)
+    }
+}
+
+impl<'a, B: Brush> From<&'a [FontFamily<'a>]> for StyleProperty<'a, B> {
+    fn from(fs: &'a [FontFamily<'a>]) -> Self {
+        StyleProperty::FontStack(fs.into())
+    }
+}
+
+impl<'a, B: Brush> From<FontFamily<'a>> for StyleProperty<'a, B> {
+    fn from(f: FontFamily<'a>) -> Self {
+        StyleProperty::FontStack(FontStack::from(f))
+    }
+}
+
+impl<B: Brush> From<GenericFamily> for StyleProperty<'_, B> {
+    fn from(f: GenericFamily) -> Self {
+        StyleProperty::FontStack(f.into())
+    }
+}
